iTrue_G is actually integer.
Former-commit-id: 3710dae57c7c97cad22ddc4c9265a2252478fa47

+++ src/ModCellGradient.f90
@@ -16,7 +16,7 @@ module ModCellGradient
   public:: calc_divergence
   public:: calc_gradient
 
-  real, allocatable:: iTrue_G(:,:,:)
+  integer, allocatable:: iTrue_G(:,:,:)
 
 contains
